dbeaver/cloudbeaver#4438 Move to SQLUtils
